backend/pyproject.toml
[project]
name = "agniraksha-backend"
version = "0.0.1"
description = "AgniRakhsa Fire Detection — Backend & AI Core"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
    "httpx>=0.27.0",
    "bcrypt>=4.2.0",
    "python-multipart>=0.0.9",
    "ultralytics>=8.1.0",
    "Pillow>=10.2.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.9.0",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
